+++ CHANGES.md
@@ -1,4 +1,5 @@
 ### Skin Composer Version 40 ###
+* Allow reordering of styles by drag and drop in the styles menu.
 * Added an option to create a single pixel texture region that will be added to the drawables.
 * Added Auto Patches functionality to Ten Patch dialog.
 * The cursor and selection in the preview of the BitmapFont/FreeTypeFont dialogs now changes color based on font color.